+++ .github/workflows/dist_build.yml
@@ -4,12 +4,69 @@ on:
   workflow_call:
     outputs:
       artifact_name:
-        value: ${{ jobs.build.outputs.artifact_name }}
+        value: ${{ jobs.collect.outputs.artifact_name }}
       run_id:
         value: ${{ github.run_id }}
 
 jobs:
-  build:
+
+  # soerp now ships a compiled FORTRAN extension, so wheels are
+  # platform- and interpreter-specific: one manylinux wheel can no longer
+  # serve every runner. cibuildwheel produces the full matrix.
+  wheels:
+    name: Wheels (${{ matrix.os }})
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      fail-fast: true
+      matrix:
+        os:
+          - ubuntu-latest
+          - windows-latest
+          # Pinned: Homebrew bottles on macos-latest carry a minos that
+          # delocate rejects against the deployment target below. Keep this
+          # image and MACOSX_DEPLOYMENT_TARGET in sync if either changes.
+          - macos-15
+
+    steps:
+      - uses: actions/checkout@v7
+        with:
+          fetch-depth: 0
+
+      - name: Setup Python
+        uses: actions/setup-python@v7
+        with:
+          python-version: "3.12"
+
+      - name: Install gfortran (macOS)
+        if: runner.os == 'macOS'
+        run: |
+          brew install gcc || brew link --overwrite gcc
+          echo "$(brew --prefix gcc)/bin" >> "$GITHUB_PATH"
+          "$(brew --prefix gcc)/bin/gfortran" --version
+
+      - name: Install Python dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install "numpy>=2.0.0" cibuildwheel
+
+      - name: Build wheels
+        run: python -m cibuildwheel --output-dir dist
+        env:
+          CIBW_ARCHS_LINUX: x86_64
+          CIBW_ARCHS_MACOS: arm64
+          CIBW_ARCHS_WINDOWS: AMD64
+          CIBW_ENVIRONMENT_MACOS: "PATH=/opt/homebrew/bin:$PATH MACOSX_DEPLOYMENT_TARGET=15.0"
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v7
+        with:
+          name: dist-wheels-${{ matrix.os }}
+          path: dist/*.whl
+          if-no-files-found: error
+
+  sdist:
+    name: Source distribution
     runs-on: ubuntu-latest
 
     steps:
@@ -20,8 +77,31 @@ jobs:
       - name: Setup uv
         uses: astral-sh/setup-uv@v7
 
-      - name: Build distribution
-        run: uv build -v --clear .
+      - name: Build sdist
+        run: uv build --sdist -v
+
+      - name: Upload sdist
+        uses: actions/upload-artifact@v7
+        with:
+          name: dist-sdist
+          path: dist/*.tar.gz
+          if-no-files-found: error
+
+  collect:
+    name: Collect distributions
+    runs-on: ubuntu-latest
+    needs: [ wheels, sdist ]
+
+    steps:
+      - name: Download every distribution
+        uses: actions/download-artifact@v8
+        with:
+          pattern: dist-*
+          merge-multiple: true
+          path: dist
+
+      - name: List distributions
+        run: ls -l dist
 
       - name: Set Artifact Name
         id: artifact-name
