+++ .sonarcloud.properties
@@ -22,4 +22,4 @@
 # its own next to the result it annotates. python:S1192 reads it as
 # duplicated string literals, which fails the quality gate without
 # describing any risk to users.
-sonar.exclusions=.github/**,tests/**,docs/**,soerp_examples.py
+sonar.exclusions=.github/**,tests/**,docs/**,bin/**,soerp_examples.py
